Tic-tac-toe Ten
Created a Genetic algorithm to automatically tune a nueral network to
serve as a hueristic for the game of tic-tac-toe ten. Due to the nature
of the game, the minimax search algorithm yielded less than ideal
results. A monte carlo tree search algorithm was implemented to great
success. The implementation is very simple and will be expanded upon
later.